% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_predictions.R
\name{calibrated_predictions}
\alias{calibrated_predictions}
\title{Calibrate machine learning predictions}
\usage{
calibrated_predictions(REGIONS, LRN_IDS, preds, calibrations)
}
\arguments{
\item{REGIONS}{\code{character}, the \code{task.id} identifiers and the list of area of study}

\item{LRN_IDS}{\code{character}, the \code{learner.id} identifiers and the list of learners}

\item{preds}{\code{list}, output from \code{get_predictions()}}

\item{calibrations}{\code{list}, output from \code{get_calibrations()}}
}
\value{
a list of calibrated predictions
}
\description{
Calibrate machine learning predictions
}
\keyword{ml-predictions}
