% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_benchmark.R
\name{compute_benchmark}
\alias{compute_benchmark}
\title{Computes the benchmark in parallel mode}
\usage{
compute_benchmark(learners, tasks, outers, prob, mes)
}
\arguments{
\item{learners}{a list of \code{mlr} learners}

\item{tasks}{a list of \code{mlr} tasks}

\item{outers}{\code{mlr} \code{resampleDesc}, outer resampling}

\item{prob}{\code{logical}, controls the type of output, if \code{TRUE} probabilities, if \code{FALSE} response}

\item{mes}{\code{mlr} list of measure to compute while tuning, the learner are tuning against the first element}
}
\value{
a list of \code{mlr} benchmark results
}
\description{
Computes the benchmark in parallel mode
}
\keyword{ml-benchmark}
