% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_benchmark.R
\name{compute_final_model}
\alias{compute_final_model}
\title{Compute the final models in sequential mode}
\usage{
compute_final_model(learners, tasks, prob, mes)
}
\arguments{
\item{learners}{a list of \code{mlr} learners}

\item{tasks}{a list of \code{mlr} tasks}

\item{prob}{\code{logical}, controls the type of output, if \code{TRUE} probabilities, if \code{FALSE} response}

\item{mes}{\code{mlr} list of measure to compute while tuning, the learner are tuning against the first element}
}
\value{
a list of model to predict with
}
\description{
Compute the final models in sequential mode
}
\keyword{ml-benchmark}
