% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_predictions.R
\name{compute_predictions}
\alias{compute_predictions}
\title{Retrieve the final (optimal) models}
\usage{
compute_predictions(REGIONS, LRN_IDS, mods, TARGET, TUNE_FS)
}
\arguments{
\item{REGIONS}{\code{character}, the \code{task.id} identifiers and the list of area of study}

\item{LRN_IDS}{\code{character}, the \code{learner.id} identifiers and the list of learners}

\item{mods}{a named list of models returned by \code{get_final_models()}}

\item{TARGET}{\code{logical}, default to \code{TRUE}, toggle the target or training datasets}

\item{TUNE_FS}{\code{logical}, toggle the behaviour for the feature selection}
}
\value{
a named list with predictions
}
\description{
Retrieve the final (optimal) models
}
\keyword{ml-predictions}
