% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_StreamCat.R
\name{drops_streamcat_df}
\alias{drops_streamcat_df}
\title{Remove some variables from \code{streamcat_df}}
\usage{
drops_streamcat_df(streamcat_df, .type = "WsCat")
}
\arguments{
\item{streamcat_df}{a \code{data.frame}}

\item{.type}{\code{character}, one of \code{WsCat}, \code{Ws} or \code{Cat}}
}
\value{
a \code{data.frame} with dropped variables
}
\description{
Remove some variables from \code{streamcat_df}
}
