% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_post_processing.R
\name{fixVarNames}
\alias{fixVarNames}
\title{Fixes variable names for data visualization purpose}
\usage{
fixVarNames(df, var = "var")
}
\arguments{
\item{df}{a \code{data.frame}}

\item{var}{\code{character} the name of the \code{var} column in `df}
}
\value{
a \code{data.frame} with fixed variable names
}
\description{
Fixes variable names for data visualization purpose
}
\keyword{ml-postprocess}
