% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_data_loading.R
\name{fmt_labels}
\alias{fmt_labels}
\title{Format the class labels}
\usage{
fmt_labels(groups, region)
}
\arguments{
\item{groups}{the class groups}

\item{region}{\code{character}, identifier of the region of study}
}
\value{
a \code{factor} with formatted groups in the format \code{regionXX}
}
\description{
Format the class labels
}
\keyword{ml-data-loading}
