% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_post_processing.R
\name{get_bestFeatureSets}
\alias{get_bestFeatureSets}
\title{Get the optinmal feature sets}
\usage{
get_bestFeatureSets(
  BMR_perf,
  selectedPATH,
  lrn = c("svm", "randomForest", "nnTrain"),
  window_size = 5,
  confidence_level = 0.05
)
}
\arguments{
\item{BMR_perf}{a \code{data.frame} from \code{getBMR_perf_tune()}}

\item{selectedPATH}{\code{character}, path to the selected features}

\item{lrn}{\code{character}, mlr \code{learner.id}}

\item{window_size}{\code{numeric}, width of the search window}

\item{confidence_level}{\code{numeric}, confidence level}
}
\value{
a \code{data.frame} with the optimal model number of features and selected features per \code{task.id} and \code{learner.id}
}
\description{
Get the optinmal feature sets
}
\keyword{ml-postprocess}
