% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_post_processing.R
\name{get_BMR}
\alias{get_BMR}
\title{Retrieve benchmark results}
\usage{
get_BMR(BMR_dir)
}
\arguments{
\item{BMR_dir}{\code{character}, base directory to look for each regional benchmark}
}
\value{
a \code{mlr} benchmark results object
}
\description{
This function is a convenience wrapper around \code{mlr::mergeBenchmarkResults()}.
}
\keyword{ml-postprocess}
