% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_predictions.R
\name{get_calibrations}
\alias{get_calibrations}
\title{Performs a posterior multinomial calibration of machine learning predictions}
\usage{
get_calibrations(REGIONS, LRN_IDS, preds, plot = FALSE)
}
\arguments{
\item{REGIONS}{\code{character}, the \code{task.id} identifiers and the list of area of study}

\item{LRN_IDS}{\code{character}, the \code{learner.id} identifiers and the list of learners}

\item{preds}{\code{list}, output from \code{get_predictions()}}

\item{plot}{\code{logical}, if \code{TRUE} displays a calibration plot}
}
\value{
the calibration models
}
\description{
Performs a posterior multinomial calibration of machine learning predictions
}
\keyword{ml-predictions}
