% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_data_loading.R
\name{get_conf_gis}
\alias{get_conf_gis}
\title{Load a SpatialLinesDataFrame containing values for valley confinement, GIS slope and RUSLE.
The SpatialDataFrame source data is the NHDv2Plus dataset which is segmented into 200-m stream intervals.}
\usage{
get_conf_gis(region, stream_order = FALSE)
}
\arguments{
\item{stream_order}{logical, if \code{TRUE} changing the file loaded for a given region}
}
\value{
a SpatialLinesDataFrame
}
\description{
Load a SpatialLinesDataFrame containing values for valley confinement, GIS slope and RUSLE.
The SpatialDataFrame source data is the NHDv2Plus dataset which is segmented into 200-m stream intervals.
}
