% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_data_loading.R
\name{get_coords}
\alias{get_coords}
\title{Retrieve the coordinates of the observations}
\usage{
get_coords(region)
}
\arguments{
\item{region}{\code{character}, identifier of the region of study}
}
\value{
the coordinates of the labelled points for a given \code{region}
}
\description{
Retrieve the coordinates of the observations
}
\keyword{ml-data-loading}
