% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_predictions.R
\name{get_entropy_df}
\alias{get_entropy_df}
\title{Formats entropy rate results}
\usage{
get_entropy_df(entropy_rate)
}
\arguments{
\item{entropy_rate}{a list of list}
}
\value{
a formatted \code{data.frame}
}
\description{
Formats entropy rate results
}
\keyword{ml-predictions}
