% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_learners.R
\name{get_final_learners}
\alias{get_final_learners}
\title{Get the final learners from the benchmark tuning results}
\usage{
get_final_learners(wrapped_learners, PATH, region, bestBMR_tune)
}
\arguments{
\item{wrapped_learners}{\code{TuneWrapper} obtained with \code{get_learners()}}

\item{PATH}{\code{file.path} to the benchmark results}

\item{region}{\code{character} identifying the region of study and the \code{mlr} task}

\item{bestBMR_tune}{\code{data.frame}, results from the benchmark}
}
\value{
a \code{mlr} learner with most frequently tuned hyper-parameters
}
\description{
Get the final learners from the benchmark tuning results
}
\keyword{ml-learners}
