% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_predictions.R
\name{get_final_models}
\alias{get_final_models}
\title{Retrieve the final (optimal) models}
\usage{
get_final_models(mod.dir, REGIONS, TUNE_FS)
}
\arguments{
\item{mod.dir}{\code{character}, the path to the directory of the model, usually set to \code{PATH}}

\item{REGIONS}{\code{character}, the \code{task.id} identifiers and the list of area of study}

\item{TUNE_FS}{\code{logical}, toggle the behaviour for the feature selection}
}
\value{
a named list of models
}
\description{
Retrieve the final (optimal) models
}
\keyword{ml-predictions}
