% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_H_rasters.R
\name{get_H_rasters}
\alias{get_H_rasters}
\title{Retrieve rasters of statistical roughness of topography}
\usage{
get_H_rasters(extdir = NULL)
}
\arguments{
\item{extdir}{default to \code{NULL}, if not \code{NULL} this indicates a \code{file.path} or \code{character} pointing to a directory external to the package. This function expects to find shapefiles named \code{scale.grd} in \code{extdir}.}
}
\value{
a named list with two elements \code{H_rasters} and \code{upper_scales}
}
\description{
Retrieve rasters of statistical roughness of topography
}
\keyword{statistical-roughness}
