% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_benchmark.R
\name{get_inner}
\alias{get_inner}
\title{Get the inner folds for the nested resampling}
\usage{
get_inner(FINAL, NU, REPS, SPCV)
}
\arguments{
\item{FINAL}{\code{logical}, selects the type of runs, if \code{FALSE} trains, else predict}

\item{NU}{\code{numeric}, number of folds for the \eqn{\nu}-fold cross-validation}

\item{REPS}{\code{numeric}, number of repetitions for the repeated \eqn{\nu}-fold cross-validation}

\item{SPCV}{\code{logical}, if \code{TRUE} activates the spatial cross-validation. This also deactivates the stratification of the folds.}
}
\value{
a \code{ResampleDesc} describing the type of resampling
}
\description{
Get the inner folds for the nested resampling
}
\keyword{ml-benchmark}
