% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial_data_loading.R
\name{get_input_data}
\alias{get_input_data}
\title{Loading input data}
\usage{
get_input_data(fpath)
}
\arguments{
\item{fpath}{a \code{file.path} to an input \code{.csv} file}
}
\value{
a \code{data.frame} with columns \code{Name}, \code{long} and \code{lat}
}
\description{
Loading input data
}
\keyword{geospatial-data}
