% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial_data_loading.R
\name{get_input_polygons}
\alias{get_input_polygons}
\title{Get the polygons for a set of locations defined by \code{input_data} and \code{n}}
\usage{
get_input_polygons(input_data, n, DEM, .dl = 25)
}
\arguments{
\item{input_data}{a \code{data.frame} with columns \code{lon} and \code{lat}}

\item{n}{numeric the indices of the rows of input_data to use}

\item{DEM}{a \code{raster} object}

\item{.dl}{a number of pixel added in all direction to the coordinates of the i-th point in order to create a square tile}
}
\value{
a list of SpatialPolygon
}
\description{
Get the polygons for a set of locations defined by \code{input_data} and \code{n}
}
