% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_learners.R
\name{get_learners}
\alias{get_learners}
\title{Get the learners}
\usage{
get_learners(
  lrn_ids,
  tuneLength = 20,
  inner = makeResampleDesc("Holdout", stratify = TRUE),
  iters = 5,
  prob = FALSE,
  smote_data,
  mes,
  info
)
}
\arguments{
\item{lrn_ids}{\code{character}, list of \code{mlr} learner idenfitiers}

\item{tuneLength}{\code{numeric}, defines the granularity of the discrete tuning grid}

\item{inner}{\code{resampleDesc} from \code{mlr}, the inner folds of the nested resampling}

\item{iters}{\code{numeric}, the number of iteration for the random discrete tuning}

\item{prob}{\code{logical}, controls the type of output, if \code{TRUE} probabilities, if \code{FALSE} response}

\item{smote_data}{a named list with two elements \code{data} and \code{labels}, if \code{smote_data} is a \code{data.frame} it is cast into the expected list format}

\item{mes}{\code{mlr} list of measure to compute while tuning, the learner are tuning against the first element}

\item{info}{\code{logical}, controls the amount of information printed when tuning}
}
\description{
This function is a wrapper around \code{get_learners_internal()}.
}
\keyword{ml-learners}
