% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial_data_loading.R
\name{get_points_from_input_data}
\alias{get_points_from_input_data}
\title{Get the SpatialPoint from input data}
\usage{
get_points_from_input_data(
  input_data,
  .crs = sp::CRS("+proj=longlat +datum=WGS84")
)
}
\arguments{
\item{input_data}{a \code{data.frame} with columns \code{lon} and \code{lat}}

\item{.crs}{a \code{sp::CRS}, usually \code{+proj=longlat +datum=WGS84}}
}
\value{
a \code{SpatialPoints} object
}
\description{
Get the SpatialPoint from input data
}
\keyword{geospatial-data}
