% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial_data_loading.R
\name{get_pol}
\alias{get_pol}
\title{Get a SpatialPolygons around a point}
\usage{
get_pol(i, pts, dl = 1000, .crs = sp::CRS("+proj=longlat +datum=WGS84"))
}
\arguments{
\item{i}{the indice of the requested points}

\item{pts}{SpatialPoints}

\item{dl}{a number of pixel added in all direction to the coordinates of the i-th point in order to create a square tile}

\item{.crs}{a \code{sp::CRS}, usually the one from the DEM to be tiled}
}
\description{
Get a SpatialPolygons around a point
}
\keyword{geospatial-data}
