% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_data_transformation.R
\name{get_ppc}
\alias{get_ppc}
\title{Get the preprocessing}
\usage{
get_ppc(sanitized_data, seed_preproc, PREPROC)
}
\arguments{
\item{sanitized_data}{a \code{data.frame}}

\item{seed_preproc}{\code{numeric} seed for \code{set.seed()} to reproduce results}

\item{PREPROC}{vector of preprocessing identifiers (e.g. \code{"scale"})}
}
\value{
a \code{caret} model object that can be used to transform the data
}
\description{
Get the preprocessing
}
\keyword{ml-data-transformation}
