% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_predictions.R
\name{get_predictions}
\alias{get_predictions}
\title{Get predictions}
\usage{
get_predictions(REGIONS, LRN_IDS, mod.dir, TARGET = TRUE, TUNE_FS)
}
\arguments{
\item{REGIONS}{\code{character}, the \code{task.id} identifiers and the list of area of study}

\item{LRN_IDS}{\code{character}, the \code{learner.id} identifiers and the list of learners}

\item{mod.dir}{\code{character}, the path to the directory of the model, usually set to \code{PATH}}

\item{TARGET}{\code{logical}, default to \code{TRUE}, toggle the target or training datasets}

\item{TUNE_FS}{\code{logical}, toggle the behaviour for the feature selection}
}
\value{
a named list of predictions
}
\description{
This function is a convenience wrapper around the \code{get_final_models()} and \code{compute_predictions()} functions.
}
\keyword{ml-predictions}
