% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_learners.R
\name{get_ps}
\alias{get_ps}
\title{Get the values for the hyper-parameter(s) set}
\usage{
get_ps(lrn.id, data, grid_resolution)
}
\arguments{
\item{lrn.id}{\code{character}, a \code{mlr} learner identifier}

\item{data}{a named list with two elements \code{data} and \code{labels}}

\item{grid_resolution}{\code{numeric}, defines the granularity of the discrete tuning grid}
}
\value{
a \code{mlr} \code{paramSet}
}
\description{
Here is the list of learner supported by this function:
\itemize{
\item "classif.svm"
\item "classif.randomForest"
\item "classif.plsdaCaret"
\item "classif.ksvm"
\item "classif.earth"
\item "classif.IBk"
\item "classif.fnn"
\item "classif.rpart"
\item "classif.lda"
\item "classif.mda"
\item "classif.h2o.glm"
\item "classif.h2o.gbm"
\item "classif.nnTrain"
\item "classif.h2o.deeplearning"
\item "classif.rda"
\item "classif.naiveBayes"
\item "classif.xgboost"
}
}
\keyword{ml-learners}
