% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_data_transformation.R
\name{get_smote_coords}
\alias{get_smote_coords}
\title{Produces noisy coordinates for SMOTE data}
\usage{
get_smote_coords(coords, labels, smote_data, seed_preproc)
}
\arguments{
\item{coords}{the coordinates of observations}

\item{labels}{\code{character} or \code{factor} the labels}

\item{smote_data}{a named list with two elements \code{data} and \code{labels}}

\item{seed_preproc}{\code{numeric} seed for \code{set.seed()} to reproduce results}
}
\description{
Produces noisy coordinates for SMOTE data
}
\keyword{This}
\keyword{`resolve_class_imbalance()`.}
\keyword{around}
\keyword{function}
\keyword{is}
\keyword{ml-data-transformation}
\keyword{wrapped}
