% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_data_transformation.R
\name{get_smote_data}
\alias{get_smote_data}
\title{Resolve the class imbalance using the SMOTE algorithm}
\usage{
get_smote_data(training_data, seed_preproc)
}
\arguments{
\item{seed_preproc}{\code{numeric} seed for \code{set.seed()} to reproduce results}

\item{data}{a named list with two elements \code{data} and \code{labels}}
}
\description{
This function is wrapped around \code{resolve_class_imbalance()}.
}
\keyword{ml-data-transformation}
