% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tam-dm.R
\name{get_stats_df}
\alias{get_stats_df}
\title{Wrapper function to calculate terrain analysis statistics for a RasterStack}
\usage{
get_stats_df(
  fun,
  fun_name,
  ls,
  dem_file,
  lines = NULL,
  stat = c("median", "mean", "min", "max", "sd", "skew")
)
}
\arguments{
\item{fun}{either \code{near_channel_stats} or \code{raster_stats}}

\item{fun_name}{a character string specifying the name of the function in the column of the resulting \code{data.frame}}

\item{ls}{a RasterStack of terrain analysis rasters}

\item{dem_file}{character, DEM filename (used for fixing names)}

\item{lines}{a SpatialLinesDataFrame}

\item{stat}{character, list of statistics to derive}
}
\value{
a \code{data.frame} with the values for the requested terrain analysis distribution metrics
}
\description{
Wrapper function to calculate terrain analysis statistics for a RasterStack
}
