% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_streamcat.R
\name{get_streamcat_df}
\alias{get_streamcat_df}
\title{Get the StreamCat data}
\usage{
get_streamcat_df(
  streamcat_dir,
  hydro_regions = c("Region18", "Region16", "Region15")
)
}
\arguments{
\item{streamcat_dir}{\code{character} or \code{file.path} to a directory containing the StreamCat data}

\item{hydro_regions}{\code{character}, target US HydroRegion}
}
\value{
a \code{data.frame} with StreamCat data
}
\description{
Get the StreamCat data
}
\keyword{streamcat}
