% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_data_loading.R
\name{get_target_data}
\alias{get_target_data}
\title{Load target features}
\usage{
get_target_data(region, extdir = NULL)
}
\arguments{
\item{region}{\code{character}, identifier of the region of study}

\item{extdir}{\code{character} or \code{file.path}, a directory in which to fetch the \code{groups} and \code{data_df} \code{csv} files}
}
\value{
\code{data_df} the target data \code{data.frame}
}
\description{
Load target features
}
\keyword{ml-data-loading}
