% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial_data_loading.R
\name{get_target_points}
\alias{get_target_points}
\title{Extracts mid-points from streamlines}
\usage{
get_target_points(
  target_streamlines,
  .crs = sp::CRS("+proj=longlat +datum=WGS84"),
  as.df = TRUE
)
}
\arguments{
\item{target_streamlines}{a SpatialLinesDataFrame}

\item{.crs}{a reference coordinate system to project the points to}

\item{as.df}{logical, default to \code{TRUE}, defines if a \code{SpatialPoints} or a \code{data.frame} object is returned}
}
\value{
a \code{data.frame} with columns \code{Name}, \code{long} and \code{lat}
}
\description{
Extracts mid-points from streamlines
}
\keyword{geospatial-data}
