% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial_data_loading.R
\name{get_target_points_df}
\alias{get_target_points_df}
\title{Extracts mid-points from streamlines}
\usage{
get_target_points_df(target_streamlines)
}
\arguments{
\item{target_streamlines}{a SpatialLinesDataFrame}
}
\value{
a data.frame
}
\description{
Extracts mid-points from streamlines
}
