% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_StreamCat.R
\name{get_target_streamcat_df}
\alias{get_target_streamcat_df}
\title{Matches \code{streamcat_df} and \code{target_streamlines} using \code{COMID}}
\usage{
get_target_streamcat_df(streamcat_df, target_streamlines)
}
\arguments{
\item{streamcat_df}{a \code{data.frame} with \code{COMID} column}

\item{target_streamlines}{a \code{SpatialLinesDataFrame} with \code{COMID} attribute}
}
\description{
Matches \code{streamcat_df} and \code{target_streamlines} using \code{COMID}
}
