% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial_data_loading.R
\name{get_target_streamlines}
\alias{get_target_streamlines}
\title{Load a SpatialLinesDataFrame containing values for valley confinement, GIS slope and RUSLE.
The SpatialDataFrame source data is the NHDv2Plus dataset which is segmented into 200-m stream intervals.}
\usage{
get_target_streamlines(region, extdir = NULL)
}
\arguments{
\item{region}{defines the area of study and the files loaded}

\item{extdir}{default to \code{NULL}, if not \code{NULL} this indicates a \code{file.path} or \code{character} pointing to a directory external to the package. This function expects to find shapefiles named \code{region_200m_VAA_LDD.shp} in \code{extdir}.}
}
\value{
a SpatialLinesDataFrame
}
\description{
Load a SpatialLinesDataFrame containing values for valley confinement, GIS slope and RUSLE.
The SpatialDataFrame source data is the NHDv2Plus dataset which is segmented into 200-m stream intervals.
}
\keyword{geospatial-data}
