% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tam-dm.R
\name{get_terrain_metrics}
\alias{get_terrain_metrics}
\title{Get the terrain analysis metrics}
\usage{
get_terrain_metrics(polys, DEM, curvature = FALSE)
}
\arguments{
\item{polys}{a list of SpatialPolygon}

\item{DEM}{a Raster}

\item{curvature}{logical, default to \code{FALSE}, if \code{TRUE} allows the calculation of profile and planform curvature}
}
\description{
Get the terrain analysis metrics
}
\keyword{tam-dm}
