% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_data_loading.R
\name{get_training_data}
\alias{get_training_data}
\title{Load training features}
\usage{
get_training_data(region, extdir = NULL)
}
\arguments{
\item{region}{\code{character}, identifier of the region of study}

\item{extdir}{\code{character} or \code{file.path}, a directory in which to fetch the \code{groups} and \code{data_df} \code{csv} files}
}
\value{
a named list with two elements \code{groups} the classes of each observation, \code{data_df} the training data \code{data.frame}
}
\description{
Load training features
}
\keyword{ml-data-loading}
