% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_post_processing.R
\name{getAllBMRS}
\alias{getAllBMRS}
\title{Get all the benchmark result in a directory of directories}
\usage{
getAllBMRS(PATH, pattern = "FS_")
}
\arguments{
\item{PATH}{initial path}

\item{pattern}{\code{character}, default to \code{FS_}, the pattern for selecting the sub-directories}
}
\value{
a named list with two elements \code{perf} and \code{tune} the benchmark performance results and the tuning results
}
\description{
This function walks the tree of folder to retrieve all the benchmark results usually organized by learners then by regions.
The function leverages \code{get_BMR()}, \code{mlr::getBMRPerformances()} and \code{mlr::getBMRTuneResults()}.
}
\keyword{ml-postprocess}
