% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_post_processing.R
\name{getBestBMRTune}
\alias{getBestBMRTune}
\title{Get the tuning results of the optimal models}
\usage{
getBestBMRTune(BMR_tune, bestFeatureSets)
}
\arguments{
\item{BMR_tune}{a \code{data.frame} containing the results of the tuning}

\item{bestFeatureSets}{the optimal selected features for each regions}
}
\value{
a \code{data.frame} subsetted to correspond to the \code{bestFeatureSets}
}
\description{
Get the tuning results of the optimal models
}
\keyword{ml-postprocess}
