% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_post_processing.R
\name{getBestBMRTuningEntropy}
\alias{getBestBMRTuningEntropy}
\title{Get the tuning entropy corresponding to the optimal models}
\usage{
getBestBMRTuningEntropy(BMR_lrnH, bestFeatureSets, bestBMR_tune)
}
\arguments{
\item{BMR_lrnH}{a \code{data.frame} generated with \code{getBMRTuningEntropy()}}

\item{bestFeatureSets}{the optimal selected features for each regions}

\item{bestBMR_tune}{a \code{data.frame} containing the results of the tuning for optimal models}
}
\description{
Get the tuning entropy corresponding to the optimal models
}
\keyword{ml-postprocess}
