% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_post_processing.R
\name{getBMR_perf_tune}
\alias{getBMR_perf_tune}
\title{Selecting elements from \code{BMR_res} and cosmetic changes}
\usage{
getBMR_perf_tune(BMR_res, type = "perf")
}
\arguments{
\item{BMR_res}{a \code{BMR_res} obtained from \code{get_BMR()} (or \code{getAllBMRS()})}

\item{type}{\code{character}, \code{perf} or \code{tune}}
}
\value{
a \code{data.frame} with the selected results
}
\description{
The cosmetic changes are simple name-fixing in the column and in the \code{task.id}.
}
\keyword{ml-postprocess}
