% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_post_processing.R
\name{getBMRTuningEntropy}
\alias{getBMRTuningEntropy}
\title{Compute the tuning entropy}
\usage{
getBMRTuningEntropy(
  BMR_tune,
  TUNELENGTH,
  maxH_nnTrain = sum(sapply(c(5, 7, 5, 3, 5, 3, 3), log2))
)
}
\arguments{
\item{BMR_tune}{a \code{data.frame} containing the results of the tuning}

\item{TUNELENGTH}{\code{numeric}, tuning length}

\item{maxH_nnTrain}{\code{numeric} the maximum value for entropy for the \code{nnTrain} learner}
}
\value{
a \code{data.frame}
}
\description{
Compute the tuning entropy
}
\keyword{ml-postprocess}
