% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_post_processing.R
\name{getFreqBestFeatureSets}
\alias{getFreqBestFeatureSets}
\title{Get the frequency of selection of a given feature across all regions}
\usage{
getFreqBestFeatureSets(bestFeatureSets)
}
\arguments{
\item{bestFeatureSets}{the optimal selected features for each regions}
}
\value{
a \code{data.frame}
}
\description{
Get the frequency of selection of a given feature across all regions
}
\keyword{ml-postprocess}
