% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_post_processing.R
\name{getHyperparNames}
\alias{getHyperparNames}
\title{Get hyper-parameters names}
\usage{
getHyperparNames(lrn)
}
\arguments{
\item{lrn}{a \code{mlr} \code{learner.id}}
}
\value{
\code{character} list of the hyper-parameter(s) name(s)
}
\description{
Get hyper-parameters names
}
\keyword{ml-postprocess}
