% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial_data_loading.R
\name{getpol}
\alias{getpol}
\title{Get a SpatialPolygons around a point}
\usage{
getpol(i, dl = 1000, .crs = sp::CRS("+proj=longlat +datum=WGS84"), .pts = pts)
}
\arguments{
\item{i}{the indice of the requested points}

\item{dl}{a number of pixel added in all direction to the coordinates of the i-th point in order to create a square tile}

\item{.crs}{a \code{sp::CRS}, usually the one from the DEM to be tiled}

\item{.pts}{SpatialPoints}
}
\description{
Get a SpatialPolygons around a point
}
