% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_post_processing.R
\name{getTunePlot}
\alias{getTunePlot}
\title{Plot the distribution of hyper-parameters resulting from the nested resampling}
\usage{
getTunePlot(BMR, lrn)
}
\arguments{
\item{BMR}{a \code{data.frame} with tuning results}

\item{lrn}{\code{character} a \code{mlr} \code{learner.id}}
}
\value{
a \code{ggplot} object
}
\description{
Plot the distribution of hyper-parameters resulting from the nested resampling
}
\keyword{ml-postprocess}
