% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPC_optim.R
\name{HPC_optim}
\alias{HPC_optim}
\title{Computes an optimized repartition of the HPC computing}
\usage{
HPC_optim(
  len,
  hpc_lim = 10000,
  io_time = 20000,
  computing_time = 1830,
  AssocGrpCpuLimit = 258,
  pl = TRUE,
  core_per_task = 1
)
}
\arguments{
\item{len}{the number of time the computation needs to be iterated}

\item{hpc_lim}{the maximum number of subjobs per user}

\item{io_time}{the input/output time in ms, default to \code{20e3}}

\item{computing_time}{the computing time in ms, default to \code{1830}}

\item{AssocGrpCpuLimit}{the maximum number of CPU per user, default to \code{258}}

\item{pl}{logical, if \code{TRUE} a plot is produced, default to \code{TRUE}}

\item{core_per_task}{number of CPU per task, default to 1.}
}
\value{
a \code{data.frame} with columns \code{array_per_task}, \code{exec_time_min} and \code{n_array}
}
\description{
This function uses the limits of maximum concurrent jobs, the input/output time and computng time.
The time are derived using \code{profVis} package.
}
