% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_H_rasters.R
\name{join_streamlines_with_H_rasters}
\alias{join_streamlines_with_H_rasters}
\title{Extract the values of all H_rasters along streamlines}
\usage{
join_streamlines_with_H_rasters(H_rasters, streamlines, upper_scales)
}
\arguments{
\item{H_rasters}{a list of Raster}

\item{streamlines}{a SpatialLinesDataFrame}

\item{upper_scales}{the upper scales corresponding to \code{H_rasters} (for naming)}
}
\value{
a \code{data.frame}
}
\description{
Extract the values of all H_rasters along streamlines
}
\keyword{statistical-roughness}
