% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_data_loading.R
\name{make_training_data}
\alias{make_training_data}
\title{Transform training data from \code{list} to \code{data.frame}}
\usage{
make_training_data(smote_data)
}
\arguments{
\item{smote_data}{a named list with two elements \code{data} and \code{labels}}
}
\value{
a \code{data.frame} with an additional column \code{channel_type}
}
\description{
Transform training data from \code{list} to \code{data.frame}
}
\keyword{ml-data-loading}
