% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_post_processing.R
\name{makeAllFeatureImportancePlotFS}
\alias{makeAllFeatureImportancePlotFS}
\title{Create feature importance across all regions of study}
\usage{
makeAllFeatureImportancePlotFS(BMR_perf, selectedPATH, bestFeatureSets)
}
\arguments{
\item{BMR_perf}{a \code{data.frame} of benchmark performance results}

\item{selectedPATH}{\code{character}, path to the selected features}

\item{bestFeatureSets}{the optimal selected features for each regions}
}
\value{
a named list of feature importance plot with names pulled from \code{task.id}
}
\description{
This function is a wrapper around \code{makeFeatureImportancePlot()} and additionally highlights which features are selected in the optimal feature sets.
}
\keyword{ml-postprocess}
