% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_post_processing.R
\name{makeAverageAccPlot}
\alias{makeAverageAccPlot}
\title{Make average accuracy plot}
\usage{
makeAverageAccPlot(BMR_perf)
}
\arguments{
\item{BMR_perf}{a \code{data.frame} of benchmark performance results}
}
\value{
a \code{ggplot} object
}
\description{
Make average accuracy plot
}
\keyword{ml-postprocess}
