% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_post_processing.R
\name{makeAverageAUCPlot}
\alias{makeAverageAUCPlot}
\title{Make average AUC plot}
\usage{
makeAverageAUCPlot(BMR_perf)
}
\arguments{
\item{BMR_perf}{a \code{data.frame} of benchmark performance results}
}
\value{
a \code{ggplot} object
}
\description{
Make average AUC plot
}
\keyword{ml-postprocess}
