% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_post_processing.R
\name{makeBestTuneAUCPlot}
\alias{makeBestTuneAUCPlot}
\title{Make a violin plot comparing the results from the optimal models}
\usage{
makeBestTuneAUCPlot(bestBMR_tune)
}
\arguments{
\item{bestBMR_tune}{a \code{data.frame} containing the results of the optimal models}
}
\value{
a \code{ggplot} object
}
\description{
The significant difference between the models are shown with a t-test with unequal variance
}
\keyword{ml-postprocess}
