% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_post_processing.R
\name{makeExampleModelSelectionPlot}
\alias{makeExampleModelSelectionPlot}
\title{Make an example plot of model selection}
\usage{
makeExampleModelSelectionPlot(
  BMR_perf,
  lrn = c("svm", "randomForest"),
  confidence_level = 0.05,
  window_size = 7
)
}
\arguments{
\item{BMR_perf}{a \code{data.frame} from \code{getBMR_perf_tune()}}

\item{lrn}{\code{character}, mlr \code{learner.id}}

\item{confidence_level}{\code{numeric}, confidence level}

\item{window_size}{\code{numeric}, width of the search window}
}
\value{
a named list with names from \code{task.id}
}
\description{
Make an example plot of model selection
}
\keyword{ml-postprocess}
