% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_post_processing.R
\name{makeFeatureImportancePlot}
\alias{makeFeatureImportancePlot}
\title{Makes a dot chart of feature importance}
\usage{
makeFeatureImportancePlot(FeatureImportance, first = 20, best = FALSE)
}
\arguments{
\item{FeatureImportance}{a \code{data.frame} with column \code{Overall} and \code{var}}

\item{first}{\code{numeric}, number of feature to display}

\item{best}{\code{logical}, default to \code{FALSE}, to highlight which features are selected in an optimal (best) set}
}
\value{
a \code{ggplot} object
}
\description{
Makes a dot chart of feature importance
}
\keyword{ml-postprocess}
