% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_post_processing.R
\name{makeTotalTimetrainPlot}
\alias{makeTotalTimetrainPlot}
\title{Make training time plot}
\usage{
makeTotalTimetrainPlot(BMR_perf)
}
\arguments{
\item{BMR_perf}{a \code{data.frame} of benchmark performance results}
}
\value{
a \code{ggplot} object
}
\description{
Make training time plot
}
\keyword{ml-postprocess}
