% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_post_processing.R
\name{makeTuningEntropyPlot}
\alias{makeTuningEntropyPlot}
\title{Create a plot of the evolution of tuning entropy with the number of selected features}
\usage{
makeTuningEntropyPlot(BMR_lrnH, bestBMR_lrnH)
}
\arguments{
\item{BMR_lrnH}{a \code{data.frame} generated with \code{getBMRTuningEntropy()}}

\item{bestBMR_lrnH}{a \code{data.frame} with the tuning entropy corresponding to the optimal models}
}
\description{
Create a plot of the evolution of tuning entropy with the number of selected features
}
\keyword{ml-postprocess}
