% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_post_processing.R
\name{makeWindowInfluencePlot}
\alias{makeWindowInfluencePlot}
\title{Visualize the influence of window size on model selection}
\usage{
makeWindowInfluencePlot(BMR_perf, selectedPATH, lrn = "randomForest")
}
\arguments{
\item{BMR_perf}{a \code{data.frame} from \code{getBMR_perf_tune()}}

\item{selectedPATH}{\code{character}, path to the selected features}

\item{lrn}{\code{character}, mlr \code{learner.id}}
}
\value{
a \code{ggplot} object
}
\description{
Visualize the influence of window size on model selection
}
\keyword{ml-postprocess}
