% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_post_processing.R
\name{normH}
\alias{normH}
\title{Calculates entropy}
\usage{
normH(var)
}
\arguments{
\item{var}{numeric, vector of values}
}
\value{
entropy in log2 units
}
\description{
Calculates entropy
}
\keyword{ml-postprocess}
