% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_data_transformation.R
\name{preproc_data}
\alias{preproc_data}
\title{Process the data according to a \code{preProcess} object}
\usage{
preproc_data(raw_training_data, ppc, labels)
}
\arguments{
\item{raw_training_data}{a \code{data.frame} of sanitized data}

\item{ppc}{a \code{preProcess} object from \code{get_ppc()}}

\item{labels}{\code{character} or \code{factor} the labels}
}
\value{
a named list with two elements: \code{data} a \code{data.frame} with the transformed data; \code{labels} the labels
}
\description{
Process the data according to a \code{preProcess} object
}
\keyword{ml-data-transformation}
