% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_characteristics.R
\docType{data}
\name{regional_characteristics}
\alias{regional_characteristics}
\title{Regional characteristics}
\format{
A \code{data.frame} with column names \code{region}, \code{n.obs}, \code{n.smote.obs}, \code{n.classes}, \code{n.conf.classes} and \code{area.sqkm}
}
\usage{
data(regional_characteristics)
}
\description{
Regional characteristics
}
\keyword{dataset}
