% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_data_transformation.R
\name{resolve_class_imbalance}
\alias{resolve_class_imbalance}
\title{Resolve class imbalances using \code{UBL} package}
\usage{
resolve_class_imbalance(data, labels, type = "Smote", .pert = 0.1)
}
\arguments{
\item{data}{a \code{data.frame}}

\item{labels}{\code{character} or \code{factor}, the labels}

\item{type}{\code{character}, a switch for the type of imbalance, one of \code{legacy_noise}, \code{Smote}, \code{GaussianNoise}, \code{Under}, \code{GaussianExtreme}}

\item{.pert}{\code{numeric}, perturbation in the Gaussian noise}
}
\description{
This function is a wrapper around \code{caret::upSample()}, \code{UBL::SmoteClassif()}, \code{UBL::GaussNoiseClassif()} and \code{UBL::RandUnderClassif()}.
}
\keyword{ml-data-transformation}
