% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_predictions.R
\name{richness}
\alias{richness}
\title{Calculate the richness (number of different species)}
\usage{
richness(l)
}
\arguments{
\item{l}{a list of values numeric or factor}
}
\value{
the richness
}
\description{
Calculate the richness (number of different species)
}
\keyword{ml-predictions}
