% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_data_loading.R
\name{sanitize_data}
\alias{sanitize_data}
\title{Ensures that non-finite values are flagged as \code{NA}}
\usage{
sanitize_data(data_df)
}
\arguments{
\item{data_df}{a \code{data.frame}}
}
\value{
a \code{data.frame}
}
\description{
Ensures that non-finite values are flagged as \code{NA}
}
\keyword{ml-data-loading}
