% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_benchmark.R
\name{savePARAMETERS}
\alias{savePARAMETERS}
\title{Saves the benchmark parameters}
\usage{
savePARAMETERS(PATH, FINAL)
}
\arguments{
\item{PATH}{path to the output directory}

\item{FINAL}{\code{logical}, selects the type of runs, if \code{FALSE} trains, else predict}
}
\description{
This function grabs the following variables from the Global Environment:
\itemize{
\item REGIONS
\item LRN_IDS
\item PREPROC
\item TUNELENGTH
\item ITERS
\item PROB
\item FINAL
\item PATH
\item NU
\item REPS
\item INNER
\item MES
\item FS
\item TUNE_FS
\item FS_NUM
}
}
\keyword{ml-benchmark}
