% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_datasets.R
\docType{data}
\name{SFE_all_data_df}
\alias{SFE_all_data_df}
\title{Target data for SFE region}
\format{
A \code{data.frame}
}
\usage{
data(SFE_all_data_df)
}
\description{
Target data for SFE region
}
\keyword{dataset}
