% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_predictions.R
\name{shannon_weiner}
\alias{shannon_weiner}
\title{Calculates Shannon-Weiner entropy}
\usage{
shannon_weiner(l, .prob_bool = FALSE, units = exp(1))
}
\arguments{
\item{l}{a list of values numeric or factor}

\item{.prob_bool}{if TRUE, l is understood as a list of probabilities, if FALSE l is understood as a factor}

\item{units}{units of the logarithm in the calculation of entropy}
}
\value{
Shannon-Weiner entropy
}
\description{
Calculates Shannon-Weiner entropy
}
\keyword{ml-predictions}
