% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_predictions.R
\name{simpson_evenness}
\alias{simpson_evenness}
\title{Calculate Simpson's evenness}
\usage{
simpson_evenness(l, .prob_bool = FALSE)
}
\arguments{
\item{l}{a list of values numeric or factor}

\item{.prob_bool}{if TRUE, l is understood as a list of probabilities, if FALSE l is understood as a factor}
}
\value{
Simpson's evenness
}
\description{
Calculate Simpson's evenness
}
\keyword{ml-predictions}
