% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial_data_loading.R
\name{snap_points_to_points}
\alias{snap_points_to_points}
\title{Snaps points to points}
\usage{
snap_points_to_points(points_a, points_b)
}
\arguments{
\item{points_a}{a \code{SpatialPoints} object}
}
\value{
\code{numeric}, the indices in \code{points_b} that are the closest to each \code{points_a}
}
\description{
Snaps points to points
}
