% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_streamlines_datasets.R
\docType{data}
\name{target_streamlines_K}
\alias{target_streamlines_K}
\title{Target streamlines for the K region}
\format{
A SpatialLinesDataFrame
}
\usage{
data(target_streamlines_K)
}
\description{
Target streamlines for the K region
}
\keyword{datasets}
