% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_streamlines_datasets.R
\docType{data}
\name{target_streamlines_NC}
\alias{target_streamlines_NC}
\title{Target streamlines for the NC region}
\format{
A SpatialLinesDataFrame
}
\usage{
data(target_streamlines_NC)
}
\description{
Target streamlines for the NC region
}
\keyword{dataset}
