% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_streamlines_datasets.R
\docType{data}
\name{target_streamlines_NCC}
\alias{target_streamlines_NCC}
\title{Target streamlines for the NCC region}
\format{
A SpatialLinesDataFrame
}
\usage{
data(target_streamlines_NCC)
}
\description{
Target streamlines for the NCC region
}
\keyword{datasets}
