% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_streamlines_datasets.R
\docType{data}
\name{target_streamlines_SAC}
\alias{target_streamlines_SAC}
\title{Target streamlines for the SAC region}
\format{
A SpatialLinesDataFrame
}
\usage{
data(target_streamlines_SAC)
}
\description{
Target streamlines for the SAC region
}
\keyword{dataset}
