% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_streamlines_datasets.R
\docType{data}
\name{target_streamlines_SC}
\alias{target_streamlines_SC}
\title{Target streamlines for the SC region}
\format{
A SpatialLinesDataFrame
}
\usage{
data(target_streamlines_SC)
}
\description{
Target streamlines for the SC region
}
\keyword{dataset}
