% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_streamlines_datasets.R
\docType{data}
\name{target_streamlines_SCC}
\alias{target_streamlines_SCC}
\title{Target streamlines for the SCC region}
\format{
A SpatialLinesDataFrame
}
\usage{
data(target_streamlines_SCC)
}
\description{
Target streamlines for the SCC region
}
\keyword{datasets}
