% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_streamlines_datasets.R
\docType{data}
\name{target_streamlines_SECA}
\alias{target_streamlines_SECA}
\title{Target streamlines for the SECA region}
\format{
A SpatialLinesDataFrame
}
\usage{
data(target_streamlines_SECA)
}
\description{
Target streamlines for the SECA region
}
\keyword{datasets}
