% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_streamlines_datasets.R
\docType{data}
\name{target_streamlines_SFE}
\alias{target_streamlines_SFE}
\title{Target streamlines for the SFE region}
\format{
A SpatialLinesDataFrame
}
\usage{
data(target_streamlines_SFE)
}
\description{
Target streamlines for the SFE region
}
\keyword{dataset}
