% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_streamlines_datasets.R
\docType{data}
\name{target_streamlines_SJT}
\alias{target_streamlines_SJT}
\title{Target streamlines for the SJT region}
\format{
A SpatialLinesDataFrame
}
\usage{
data(target_streamlines_SJT)
}
\description{
Target streamlines for the SJT region
}
\keyword{dataset}
