% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_unix.R
\name{terrain_}
\alias{terrain_}
\title{Modified \code{terrain()} function}
\usage{
terrain_(x, opt = "slope", unit = "radians", neighbors = 8, filename = "", ...)
}
\arguments{
\item{x}{raster}

\item{opt}{type of metrics}

\item{unit}{radians or tangent}

\item{neighbors}{default to queen case}

\item{filename}{passed to WriteRaster}
}
\value{
a raster with the computed metrics
}
\description{
This modified version of \code{raster::terrain()} can compute curvature. See details.
}
\details{
Author: Robert J. Hijmans
Date : February 2011
Version 1.0
Licence GPL v3
Modified by Herve Guillon, November 2017
Added options for computation of planform and profile curvature
}
\keyword{tam-dm}
