% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_benchmark.R
\name{savePARAMETERS}
\alias{savePARAMETERS}
\title{Saves the benchmark parameters}
\usage{
savePARAMETERS(PATH, FINAL)
}
\arguments{
\item{PATH}{path to the output directory}

\item{FINAL}{\code{logical}, selects the type of runs, if \code{FALSE} trains, else predict}
}
\description{
Saves the benchmark parameters
}
\keyword{ml-benchmark}
